+++ samples/AgentWarmupE2E/Tests/MongoDbAgentIdentifierServiceTests.cs
