+++ samples/AgentWarmupE2E/Utilities/PerformanceMonitor.cs
