+++ samples/AgentWarmupE2E/Utilities/TestDataGenerator.cs
