+++ src/Aevatar.Silo/AgentWarmup/DESIGN.md
