+++ src/Aevatar.Silo/AgentWarmup/IAgentWarmupStrategy.cs
