+++ test/Aevatar.Silo.Tests/AgentWarmup/TestGrains/TestWarmupAgent.cs
@@ -1,21 +1,25 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Orleans;
 using Orleans.Runtime;
 
-namespace E2E.Grains;
+namespace Aevatar.Silo.Tests.AgentWarmup.TestAgents;
 
 /// <summary>
-/// Test grain implementation for warmup validation
+/// Test agent implementation for warmup validation
 /// Tracks activation time and access patterns
 /// </summary>
-public class TestWarmupGrain : Grain, ITestWarmupGrain
+[GrainType("silo-test-warmup-agent")]
+public class TestWarmupAgent : Grain, ITestWarmupAgent
 {
-    private readonly ILogger<TestWarmupGrain> _logger;
+    private readonly ILogger<TestWarmupAgent> _logger;
     private DateTime _activationTime;
     private int _accessCount;
     private bool _isWarmedUp;
 
-    public TestWarmupGrain(ILogger<TestWarmupGrain> logger)
+    public TestWarmupAgent(ILogger<TestWarmupAgent> logger)
     {
         _logger = logger;
     }
@@ -26,7 +30,7 @@ public class TestWarmupGrain : Grain, ITestWarmupGrain
         _accessCount = 0;
         _isWarmedUp = false;
         
-        _logger.LogInformation("TestWarmupGrain {GrainId} activated at {ActivationTime}", 
+        _logger.LogInformation("TestWarmupAgent {AgentId} activated at {ActivationTime}", 
             this.GetPrimaryKey(), _activationTime);
         
         return base.OnActivateAsync(cancellationToken);
@@ -35,9 +39,9 @@ public class TestWarmupGrain : Grain, ITestWarmupGrain
     public Task<string> PingAsync()
     {
         _accessCount++;
-        var response = $"Pong from grain {this.GetPrimaryKey()} at {DateTime.UtcNow:HH:mm:ss.fff}";
+        var response = $"Pong from agent {this.GetPrimaryKey()} at {DateTime.UtcNow:HH:mm:ss.fff}";
         
-        _logger.LogDebug("Ping received by grain {GrainId}, access count: {AccessCount}", 
+        _logger.LogDebug("Ping received by agent {AgentId}, access count: {AccessCount}", 
             this.GetPrimaryKey(), _accessCount);
         
         return Task.FromResult(response);
@@ -56,7 +60,7 @@ public class TestWarmupGrain : Grain, ITestWarmupGrain
         // Simulate some computation work
         var result = input * 2 + _accessCount;
         
-        _logger.LogDebug("Computation performed by grain {GrainId}: {Input} -> {Result}", 
+        _logger.LogDebug("Computation performed by agent {AgentId}: {Input} -> {Result}", 
             this.GetPrimaryKey(), input, result);
         
         return Task.FromResult(result);
@@ -71,25 +75,25 @@ public class TestWarmupGrain : Grain, ITestWarmupGrain
     {
         _accessCount++;
         
-        _logger.LogDebug("Simulating database operation for grain {GrainId} with {DelayMs}ms delay", 
+        _logger.LogDebug("Simulating database operation for agent {AgentId} with {DelayMs}ms delay", 
             this.GetPrimaryKey(), delayMs);
         
         // Simulate database operation delay
         await Task.Delay(delayMs);
         
-        return $"Database operation completed for grain {this.GetPrimaryKey()} after {delayMs}ms";
+        return $"Database operation completed for agent {this.GetPrimaryKey()} after {delayMs}ms";
     }
 
-    public Task<GrainMetadata> GetMetadataAsync()
+    public Task<AgentMetadata> GetMetadataAsync()
     {
         _accessCount++;
         
         // Use a simpler approach for silo address - just use a placeholder for now
         string siloAddress = "TestSilo";
         
-        var metadata = new GrainMetadata
+        var metadata = new AgentMetadata
         {
-            GrainId = this.GetPrimaryKey(),
+            AgentId = this.GetPrimaryKey(),
             ActivationTime = _activationTime,
             AccessCount = _accessCount,
             SiloAddress = siloAddress,
@@ -100,19 +104,19 @@ public class TestWarmupGrain : Grain, ITestWarmupGrain
     }
 
     /// <summary>
-    /// Marks this grain as warmed up (called during warmup process)
+    /// Marks this agent as warmed up (called during warmup process)
     /// </summary>
     public void MarkAsWarmedUp()
     {
         _isWarmedUp = true;
-        _logger.LogDebug("Grain {GrainId} marked as warmed up", this.GetPrimaryKey());
+        _logger.LogDebug("Agent {AgentId} marked as warmed up", this.GetPrimaryKey());
     }
 
     public override Task OnDeactivateAsync(DeactivationReason reason, CancellationToken cancellationToken)
     {
         var activeDuration = DateTime.UtcNow - _activationTime;
         
-        _logger.LogInformation("TestWarmupGrain {GrainId} deactivating after {Duration}ms, " +
+        _logger.LogInformation("TestWarmupAgent {AgentId} deactivating after {Duration}ms, " +
                               "access count: {AccessCount}, reason: {Reason}", 
             this.GetPrimaryKey(), activeDuration.TotalMilliseconds, _accessCount, reason);
